requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
pydantic==2.8.2
PyYAML==6.0.2
markdown-it-py==3.0.0
aiosqlite==0.20.0

